New folder: addressComponents - Added the classes Block, Street, Unit, PostalCode into the folder Address.java: Removed Magic Numbers and edited comments Block.java, PostalCode.java, Unit.java, Street.java: Added comments.
