--- src/seedu/addressbook/data/person/Block.java
@@ -1,14 +1,0 @@
-package seedu.addressbook.data.person;
-
-public class Block {
-
-    public final String blockNumber;
-
-    public Block(String blockNum) {
-        this.blockNumber = blockNum;
-    }
-
-    public String getBlockNumber() {
-        return this.blockNumber;
-    }
-}
