--- src/seedu/addressbook/data/person/PostalCode.java
@@ -1,13 +1,0 @@
-package seedu.addressbook.data.person;
-
-public class PostalCode {
-    public final String postalCode;
-
-    public PostalCode(String postalCode) {
-        this.postalCode = postalCode;
-    }
-
-    public String getPostalCode() {
-        return this.postalCode;
-    }
-}
