--- src/seedu/addressbook/data/person/Street.java
@@ -1,14 +1,0 @@
-package seedu.addressbook.data.person;
-
-public class Street {
-
-    public final String streetName;
-
-    public Street(String streetNum) {
-        this.streetName = streetNum;
-    }
-
-    public String getStreetNumber() {
-        return this.streetName;
-    }
-}
