--- src/seedu/addressbook/data/person/Unit.java
@@ -1,13 +1,0 @@
-package seedu.addressbook.data.person;
-
-public class Unit {
-    public final String unitNumber;
-
-    public Unit(String unitNum) {
-        this.unitNumber = unitNum;
-    }
-
-    public String getUnitNumber() {
-        return this.unitNumber;
-    }
-}
